+++ crates/config-models/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod duration_profile;
+pub mod liquidity_profile;
 pub mod models;
 pub use models::*;
 pub mod strategy;